Created tests for engine.view.EditableElement. Moved tests from RootEditableElement to EditableElement.
