+++ tests/view/rooteditableelement.js
@@ -38,76 +38,4 @@ describe( 'RootEditableElement', () => {
 			expect( root.isReadOnly ).to.be.false;
 		} );
 	} );
-
-	describe( 'isFocused', () => {
-		let docMock, viewMain, viewHeader;
-
-		beforeEach( () => {
-			docMock = createDocumentMock();
-
-			viewMain = new RootEditableElement( docMock, 'div' );
-			viewHeader = new RootEditableElement( docMock, 'h1', 'header' );
-		} );
-
-		it( 'should be observable', () => {
-			const root = new RootEditableElement( createDocumentMock(), 'div' );
-
-			expect( root.isFocused ).to.be.false;
-
-			const isFocusedSpy = sinon.spy();
-
-			root.on( 'change:isFocused', isFocusedSpy );
-
-			root.isFocused = true;
-
-			expect( root.isFocused ).to.be.true;
-
-			expect( isFocusedSpy.calledOnce ).to.be.true;
-		} );
-
-		it( 'should change isFocused when selectedEditable changes', () => {
-			docMock.selectedEditable = viewMain;
-			docMock.isFocused = true;
-
-			expect( viewMain.isFocused ).to.be.true;
-			expect( viewHeader.isFocused ).to.be.false;
-
-			docMock.selectedEditable = viewHeader;
-
-			expect( viewMain.isFocused ).to.be.false;
-			expect( viewHeader.isFocused ).to.be.true;
-
-			docMock.selectedEditable = null;
-
-			expect( viewMain.isFocused ).to.be.false;
-			expect( viewHeader.isFocused ).to.be.false;
-		} );
-	} );
-
-	describe( 'isReadOnly', () => {
-		it( 'should be observable', () => {
-			const root = new RootEditableElement( createDocumentMock(), 'div' );
-
-			expect( root.isReadOnly ).to.be.false;
-
-			const isReadOnlySpy = sinon.spy();
-
-			root.on( 'change:isReadOnly', isReadOnlySpy );
-
-			root.isReadOnly = true;
-
-			expect( root.isReadOnly ).to.be.true;
-
-			expect( isReadOnlySpy.calledOnce ).to.be.true;
-		} );
-	} );
-
-	describe( 'getDocument', ()=> {
-		it( 'should return document', () => {
-			const docMock = createDocumentMock();
-			const root = new RootEditableElement( docMock, 'div' );
-
-			expect( root.getDocument() ).to.equal( docMock );
-		} );
-	} );
 } );
